requirements.txt
torch
torchvision
timm
numpy
scikit-learn
matplotlib
tqdm
joblib
Pillow
